Added code to extract class and interface information
Added code to parse and extract class attributes, methods, extends list and implements list to be used to create the uml grammar.
Added three new classes - ClassInterfaceAttributeInfo to store the attributes information, ClassInterfaceMethodInfo to store the methods information, and ClassInterfaceInfo to store the attribute and methods information for a class or interface.